# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SoulCourse API
  version: 0.1.0-beta
  summary: 选科π当前 API 契约骨架
  description: |
    该契约以当前 `/api/v1` 实现为基线，统一成功响应为 `{ data, meta? }`，
    错误响应为 `{ error }`，并在成功/失败中返回 `requestId`。
servers:
  - url: http://localhost:1309{basePath}/api/v1
    description: Local development
    variables:
      basePath:
        default: ""
        description: Optional APP_BASE_PATH, for example `/subject314`
security:
  - cookieAuth: []
...
paths:
  /auth/email-verification-code:
    post:
      tags: [Auth]
      operationId: sendEmailVerificationCode
      summary: 发送邮箱验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationCodeInput'
      responses:
        '200':
          $ref: '#/components/responses/EmailVerificationCodeResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: 发送重置密码验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationCodeInput'
      responses:
        '200':
          $ref: '#/components/responses/EmailVerificationCodeResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: 注册账号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          $ref: '#/components/responses/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 登录
      description: 登录成功后设置 HttpOnly、SameSite=Lax 的服务端会话 Cookie；前端不接收或存储 Bearer/JWT token。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          $ref: '#/components/responses/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /mobile/auth/register:
    post:
      tags: [Auth]
      operationId: mobileRegister
      summary: Android 客户端注册
      description: 返回可撤销的不透明移动会话令牌；网页注册仍只设置 HttpOnly Cookie。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          $ref: '#/components/responses/MobileAuthSessionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /mobile/auth/login:
    post:
      tags: [Auth]
      operationId: mobileLogin
      summary: Android 客户端登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          $ref: '#/components/responses/MobileAuthSessionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /mobile/auth/refresh:
    post:
      tags: [Auth]
      operationId: mobileRefresh
      summary: 刷新 Android 客户端会话
      security:
        - mobileBearer: []
      responses:
        '200':
          $ref: '#/components/responses/MobileAuthSessionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: 使用邮箱验证码重置密码
      description: 重置成功后撤销该账号全部服务端会话，并清理当前浏览器 Cookie。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordInput'
      responses:
        '200':
          $ref: '#/components/responses/AdminImageUploadResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 退出登录
      description: 撤销当前服务端会话并清理会话 Cookie；Cookie 会话写请求需带 X-CSRF-Token。
      security:
        - cookieAuth: []
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/LogoutResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /me:
    get:
      tags: [Profile]
      operationId: getCurrentUser
      summary: 获取当前登录用户
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Profile]
      operationId: deleteMe
      summary: 注销当前账号
      description: 校验当前密码后软删除账号、清除邮箱和密码哈希、撤销全部会话，并清理当前浏览器 Cookie。
      security:
        - cookieAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountInput'
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /me/profile:
    get:
      tags: [Profile]
      operationId: getMyProfile
      summary: 获取当前用户画像
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/AccountProfileResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    put:
      tags: [Profile]
      operationId: updateMyProfile
      summary: 更新当前用户画像
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInput'
      responses:
        '200':
          $ref: '#/components/responses/AccountProfileResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /me/sessions:
    get:
      tags: [Profile]
      operationId: listMySessions
      summary: 获取当前账号会话列表
      security:
        - cookieAuth: []
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/AccountSessionsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /me/sessions/{id}:
    delete:
      tags: [Profile]
      operationId: revokeMySession
      summary: 撤销当前账号指定会话
      security:
        - cookieAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/NumericId'
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /uploads/images/presign:
    post:
      tags: [Uploads]
      operationId: presignImageUpload
      summary: 创建图片上传会话
      security:
        - cookieAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignImageUploadInput'
      responses:
        '201':
          $ref: '#/components/responses/PresignedImageUploadResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /uploads/images/{id}/object:
    put:
      tags: [Uploads]
      operationId: putImageUploadObject
      summary: 上传图片对象
      security:
        - cookieAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UploadID'
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/gif:
            schema:
              type: string
              format: binary
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /uploads/images/{id}/complete:
    post:
      tags: [Uploads]
      operationId: completeImageUpload
      summary: 完成图片上传
      description: 服务端校验对象实际存在、归属、大小、MIME、扩展名和像素尺寸后返回可用于帖子图片的 URL。相同用户对已完成上传以相同对象元数据重试时，幂等返回原成功结果。
      security:
        - cookieAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UploadID'
      responses:
        '200':
          $ref: '#/components/responses/CompleteImageUploadResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /profiles/{name}:
    get:
      tags: [Profile]
      operationId: getProfile
      summary: 按用户名获取画像
      parameters:
        - $ref: '#/components/parameters/ProfileName'
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          $ref: '#/components/responses/AccountProfileResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      summary: 列出通知
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          $ref: '#/components/responses/NotificationsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /notifications/read-all:
    post:
      tags: [Notifications]
      operationId: markAllNotificationsRead
      summary: 全部通知标记已读
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/BooleanActionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /notifications/{id}/read:
    post:
      tags: [Notifications]
      operationId: markNotificationRead
      summary: 标记单条通知已读
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/NumericId'
      responses:
        '200':
          $ref: '#/components/responses/BooleanActionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /messages:
    get:
      tags: [Messages]
      operationId: listConversations
      summary: 列出私信会话
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationPageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          $ref: '#/components/responses/ConversationsResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Messages]
      operationId: sendDirectMessage
      summary: 发送私信
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageInput'
      responses:
        '201':
          $ref: '#/components/responses/DirectMessageResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /messages/{name}:
    get:
      tags: [Messages]
      operationId: listDirectMessages
      summary: 获取与某用户的私信
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProfileName'
      responses:
        '200':
          $ref: '#/components/responses/DirectMessagesResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
...
  /content:
    get:
      tags: [Content]
      operationId: listPublishedContent
      summary: 获取公开 CMS 内容
      parameters:
        - in: query
          name: module
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/PublishedContentResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /telemetry/web-vitals:
    post:
      tags: [Telemetry]
      operationId: reportWebVital
      summary: 匿名上报浏览器核心性能指标
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebVitalInput'
      responses:
        '204':
          description: 指标已记录
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /telemetry/mobile:
    post:
      tags: [Telemetry]
      operationId: reportMobileTelemetry
      summary: 在用户同意后匿名上报 Android 诊断事件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTelemetryInput'
      responses:
        '204':
          description: 诊断事件已记录
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /provinces:
    get:
      tags: [RealData]
      operationId: listProvinces
      summary: 列出省份数据覆盖状态
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /policies:
    get:
      tags: [RealData]
      operationId: listPolicies
      summary: 列出已复核政策内容
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /requirements:
    get:
      tags: [RealData]
      operationId: listRequirements
      summary: 列出专业选科要求内容
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /sources/{id}:
    get:
      tags: [RealData]
      operationId: getSource
      summary: 获取数据来源追溯信息
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /insights:
    get:
      tags: [Insights]
      operationId: listInsights
      summary: 列出洞察
      responses:
        '200':
          $ref: '#/components/responses/InsightsResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /insights/{id}:
    get:
      tags: [Insights]
      operationId: getInsight
      summary: 获取洞察详情
      parameters:
        - $ref: '#/components/parameters/NumericId'
      responses:
        '200':
          $ref: '#/components/responses/InsightResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /topics:
    get:
      tags: [Topics]
      operationId: listTopics
      summary: 列出话题
      responses:
        '200':
          $ref: '#/components/responses/TopicsResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /topics/{slug}:
    get:
      tags: [Topics]
      operationId: getTopic
      summary: 获取话题详情
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/TopicResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /posts/{id}/comments:
    post:
      tags: [Posts]
      operationId: createComment
      summary: 发表评论
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/NumericId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentInput'
      responses:
        '201':
          $ref: '#/components/responses/CommentResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /posts/{id}/like:
    post:
      tags: [Posts]
      operationId: togglePostLike
      summary: 点赞或取消点赞
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/NumericId'
      responses:
        '200':
          $ref: '#/components/responses/ToggleResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
...
  /admin/email-config:
    get:
      tags: [Admin]
      operationId: getAdminEmailConfig
      summary: 获取邮件配置状态
      security:
        - adminSession: []
      responses:
        '200':
          $ref: '#/components/responses/GenericObjectResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Error envelope
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    EmailVerificationCodeResponse:
      description: Email verification status
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmailVerificationCodeEnvelope'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: scf_session
      description: HttpOnly SameSite=Lax server-side session cookie. Unsafe methods also require X-CSRF-Token.
    adminSession:
      type: apiKey
      name: scf_admin_session
      in: cookie
      description: HttpOnly SameSite=Lax admin session cookie. Unsafe methods also require X-CSRF-Token with scf_admin_csrf.
    mobileBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Android 客户端使用的可撤销不透明会话令牌。
  schemas:
...
    EmailVerificationCodeInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...